feat: build script verifies versions align
The `build.gradle.kts`, `Mathlingua.kt`, and `Main.go` files
need to each specify the version number for their own uses.
The `bin/build-release.sh` script makes sure all specify the
same version.

+++ bin/build-release.sh
@@ -5,6 +5,34 @@ set -e
 
 MLG_VERSION="0.11.0"
 
+echo "Building release ${MLG_VERSION}"
+
+RED_ERROR="\033[0;31mERROR\033[0m"
+
+GRADLE_VERSION_FILE=build.gradle.kts
+echo "Checking version declaration in ${GRADLE_VERSION_FILE}"
+if [[ "$(grep "version = \"${MLG_VERSION}\"" "${GRADLE_VERSION_FILE}")" == "" ]]
+then
+  echo -e "${RED_ERROR}: ${GRADLE_VERSION_FILE} doesn't have the version set to ${MLG_VERSION}"
+  exit 1
+fi
+
+KOTLIN_VERSION_FILE=src/main/kotlin/mathlingua/cli/Mathlingua.kt
+echo "Checking version declaration in ${KOTLIN_VERSION_FILE}"
+if [[ "$(grep "const val MATHLINGUA_VERSION = \"${MLG_VERSION}\"" "${KOTLIN_VERSION_FILE}")" == "" ]]
+then
+  echo -e "${RED_ERROR}: ${KOTLIN_VERSION_FILE} doesn't have the version set to ${MLG_VERSION}"
+  exit 1
+fi
+
+GO_VERSION_FILE=mlg/Main.go
+echo "Checking version declaration in ${GO_VERSION_FILE}"
+if [[ "$(grep "const MATHLINGUA_VERSION = \"${MLG_VERSION}\"" "${GO_VERSION_FILE}")" == "" ]]
+then
+  echo -e "${RED_ERROR}: ${GO_VERSION_FILE} doesn't have the version set to ${MLG_VERSION}"
+  exit 1
+fi
+
 echo "Removing old release"
 rm -Rf release
 mkdir -p release
